Update App.js
Nice looking project. I like the design of the Home page very clean and clear design. One minnor remark would be in your components folder all your JS files are in the same folder, Its usually a good idea to create a separate folder for each component and there css so the code is a little more organized. But overall very cool project.